Fix bug when adding a new chip
Recent changes to the layout which held the chip views as children
caused the acivity to crash. The original layout was a LinearLayout
which changed to a ScrollView.
ScrollViews can only hold one direct child. This change adds the chips
as children of the internal linear layout rather than adding it to the
ScrollView.